Remove old HotkeyControl, rename HotkeyControl2 to HotkeyControl
