--- Flow.Launcher/HotkeyControl2.xaml
@@ -1,77 +1,0 @@
-﻿<UserControl
-    x:Class="Flow.Launcher.HotkeyControl2"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-    <Button
-        Width="Auto"
-        FontSize="13"
-        FontWeight="Bold"
-        Foreground="{DynamicResource Color01B}"
-        Click="GetNewHotkey">
-        <Button.Template>
-            <ControlTemplate TargetType="Button">
-                <Border
-                    x:Name="ButtonBorder"
-                    Padding="5,0,5,0"
-                    Background="{DynamicResource ButtonBackgroundColor}"
-                    BorderBrush="{DynamicResource ButtonInsideBorder}"
-                    BorderThickness="1"
-                    CornerRadius="5">
-                    <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center" />
-                </Border>
-                <ControlTemplate.Triggers>
-                    <MultiTrigger>
-                        <MultiTrigger.Conditions>
-                            <Condition Property="IsMouseOver" Value="True" />
-                            <Condition Property="IsPressed" Value="True" />
-                        </MultiTrigger.Conditions>
-                        <Setter TargetName="ButtonBorder" Property="Background"
-                                Value="{DynamicResource ButtonMousePressed}" />
-                        <Setter TargetName="ButtonBorder" Property="BorderBrush"
-                                Value="{DynamicResource ButtonMousePressedInsideBorder}" />
-                    </MultiTrigger>
-                    <MultiTrigger>
-                        <MultiTrigger.Conditions>
-                            <Condition Property="IsMouseOver" Value="True" />
-                        </MultiTrigger.Conditions>
-                        <Setter TargetName="ButtonBorder" Property="Background"
-                                Value="{DynamicResource ButtonMouseOver}" />
-                    </MultiTrigger>
-                    <MultiTrigger>
-                        <MultiTrigger.Conditions>
-                            <Condition Property="IsPressed" Value="True" />
-                        </MultiTrigger.Conditions>
-                        <Setter TargetName="ButtonBorder" Property="Background"
-                                Value="{DynamicResource ButtonMousePressed}" />
-                        <Setter TargetName="ButtonBorder" Property="BorderBrush"
-                                Value="{DynamicResource CustomContextHover}" />
-                    </MultiTrigger>
-                </ControlTemplate.Triggers>
-            </ControlTemplate>
-        </Button.Template>
-        <Button.Content>
-            <ItemsControl x:Name="HotkeyList">
-                <ItemsControl.ItemsPanel>
-                    <ItemsPanelTemplate>
-                        <StackPanel Orientation="Horizontal" />
-                    </ItemsPanelTemplate>
-                </ItemsControl.ItemsPanel>
-                <ItemsControl.ItemTemplate>
-                    <DataTemplate>
-                        <Border
-                            Margin="2,5,2,5"
-                            Padding="10,5,10,5"
-                            Background="{DynamicResource AccentButtonBackground}"
-                            BorderThickness="1"
-                            CornerRadius="5">
-                            <TextBlock Text="{Binding}" />
-                        </Border>
-                    </DataTemplate>
-                </ItemsControl.ItemTemplate>
-            </ItemsControl>
-        </Button.Content>
-    </Button>
-</UserControl>
