+++ Flow.Launcher/HotkeyControlDialog.xaml
@@ -1,5 +1,5 @@
 ﻿<ui:ContentDialog
-    x:Class="Flow.Launcher.HotkeyControl2Dialog"
+    x:Class="Flow.Launcher.HotkeyControlDialog"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:ui="http://schemas.modernwpf.com/2019"
